+++ RoboPet/RBPMiniGameScene_Music.m
@@ -482,10 +482,10 @@ typedef NS_ENUM(NSInteger, GameStateType){
 - (NSString *)gameOverMessage
 {
     if (self.score > 0.0 && self.score >= self.highScore) {
-        return [NSString stringWithFormat:@"NEW HIGH SCORE!!!\n\n%.1f", self.score];
+        return [NSString stringWithFormat:@"NEW HIGH SCORE!!!\n\n%lu", (NSInteger)self.score];
     }
     
-    return [NSString stringWithFormat:@"Score:%.1f\nHigh Score:%.1f", self.score, self.highScore];
+    return [NSString stringWithFormat:@"Score:%lu\nHigh Score:%lu", (NSInteger)self.score, (NSInteger)self.highScore];
 }
 
 @end
